+++ crates/orchestrator/src/tests/jobs/da_job/mod.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 
 use assert_matches::assert_matches;
+use chrono::{SubsecRound, Utc};
 use color_eyre::eyre::eyre;
 use mockall::predicate::always;
 use rstest::rstest;
@@ -75,6 +76,8 @@ async fn test_da_job_process_job_failure_on_small_blob_size(
                 external_id: ExternalId::String(internal_id.to_string().into_boxed_str()),
                 metadata: HashMap::default(),
                 version: 0,
+                created_at: Utc::now().round_subsecs(0),
+                updated_at: Utc::now().round_subsecs(0),
             },
         )
         .await;
@@ -137,6 +140,8 @@ async fn test_da_job_process_job_failure_on_pending_block() {
                 external_id: ExternalId::String("1".to_string().into_boxed_str()),
                 metadata: HashMap::default(),
                 version: 0,
+                created_at: Utc::now().round_subsecs(0),
+                updated_at: Utc::now().round_subsecs(0),
             },
         )
         .await;
@@ -218,6 +223,8 @@ async fn test_da_job_process_job_success(
                 external_id: ExternalId::String(internal_id.to_string().into_boxed_str()),
                 metadata: HashMap::default(),
                 version: 0,
+                created_at: Utc::now().round_subsecs(0),
+                updated_at: Utc::now().round_subsecs(0),
             },
         )
         .await;